Added policy for user, fixed authorization bugs

+++ modules/AmirVahedix/Authorization/Http/Controllers/AuthorizationController.php
@@ -6,16 +6,13 @@ namespace AmirVahedix\Authorization\Http\Controllers;
 
 use AmirVahedix\Authorization\Http\Requests\CreateRoleRequest;
 use AmirVahedix\Authorization\Http\Requests\EditRoleRequest;
+use AmirVahedix\Authorization\Models\Role;
 use AmirVahedix\Authorization\Repositories\AuthorizationRepo;
 use App\Http\Controllers\Controller;
 use Spatie\Permission\Models\Permission;
-use Spatie\Permission\Models\Role;
 
 class AuthorizationController extends Controller
 {
-    /**
-     * @var AuthorizationRepo
-     */
     private $authorizationRepo;
 
     public function __construct(AuthorizationRepo $authorizationRepo)
@@ -25,6 +22,8 @@ class AuthorizationController extends Controller
 
     public function index()
     {
+        $this->authorize('manag_roles', Role::class);
+
         $roles = Role::all();
         $permissions = Permission::all();
         return view('Authorization::index', compact('roles', 'permissions'));
@@ -32,6 +31,8 @@ class AuthorizationController extends Controller
 
     public function store(CreateRoleRequest $request)
     {
+        $this->authorize('manag_roles', Role::class);
+
         $this->authorizationRepo->createRole($request->get('name'), $request->get('permissions'));
 
         toast('نقش کاربری باموفقیت ایجاد شد.', 'success');
@@ -40,12 +41,16 @@ class AuthorizationController extends Controller
 
     public function edit(Role $role)
     {
+        $this->authorize('manag_roles', Role::class);
+
         $permissions = Permission::all();
         return view('Authorization::edit', compact('role', 'permissions'));
     }
 
     public function update(Role $role, EditRoleRequest $request)
     {
+        $this->authorize('manag_roles', Role::class);
+
         $this->authorizationRepo
             ->update($role, $request->get('name'), $request->get('permissions'));
 
@@ -55,6 +60,8 @@ class AuthorizationController extends Controller
 
     public function delete(Role $role)
     {
+        $this->authorize('manag_roles', Role::class);
+
         $role->delete();
 
         toast('نقش باموفقیت حذف شد.', 'success');
